Extarct tab panels
